node: renamed preprocess category with augment, removed incorrect img return in yolo diagram
