+++ peekingduck/pipeline/nodes/augment/brightness.py
